Add likes subscription
This commit add the like subscription and also
fixes the comment query to work with the new view
which contains the details of a post.

The comment type in the schema has also been altered to
contain the details of the author of the post and the
details of the author of the comment as well.